backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.2.1
django-ninja==1.4.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
python-dotenv==1.1.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
